+++ src/data_sources/csv.rs
@@ -17,15 +17,18 @@ pub fn write_csv(matches: &ArgMatches<'_>, df: DataFrame) -> Result<(), anyhow::
     let now_utc: chrono::DateTime<chrono::Utc> = now.into();
     let current_time = now_utc.to_rfc3339_opts(chrono::SecondsFormat::Secs, true);
 
-    let target_dir = match matches.value_of(arg::OUTPUT_DIR) {
-        Some(output_dir) => PathBuf::from(output_dir),
-        None => return Err(anyhow!("Missing value for argument '{}'", arg::OUTPUT_DIR)),
+    let target_dir = match matches.value_of(arg::SAVE_DIR) {
+        Some(save_dir) => PathBuf::from(save_dir),
+        None => return Err(anyhow!("Missing value for argument '{}'", arg::SAVE_DIR)),
     };
     let target_file = "query_".to_string() + &current_time + ".csv";
 
     match target_dir.exists() {
         true => (),
-        false => fs::create_dir(&target_dir).context("Can't create output directory")?,
+        false => fs::create_dir(&target_dir).context(format!(
+            "Can't create directory: '{}'",
+            target_dir.display()
+        ))?,
     }
 
     let target_path = target_dir.join(target_file);