+++ src/email_builder/bulk_email.rs
@@ -1,9 +1,8 @@
+use super::Mime;
 use crate::{
     arg, cmd,
     data_sources::{query_postgres, read_csv},
-    email_handler::{Confirmed, Email, Message, MessageTemplate},
-    email_provider,
-    helper::{check_send_status, format_green},
+    email_builder::{Confirmed, Email, Message, MessageTemplate},
 };
 use anyhow::{anyhow, Context, Result};
 use clap::{ArgMatches, Values};
@@ -64,27 +63,17 @@ impl BulkEmail {
         Ok(BulkEmail { emails })
     }
 
-    pub fn send(&self, matches: &ArgMatches<'_>) -> Result<(), anyhow::Error> {
+    pub fn process(&self, matches: &ArgMatches<'_>) -> Result<(), anyhow::Error> {
         println!("Sending email to {} receivers ...", self.emails.len());
 
+        for email in &self.emails {
+            email.send(matches)?;
+            email.archive(matches)?;
+        }
+
         if matches.is_present(arg::DRY_RUN) {
-            // Setup client but do not send email
-            let _client = email_provider::setup_ses_client(matches)?;
-
-            for email in &self.emails {
-                println!("{} ... {}", email.receiver, format_green("dry run"));
-            }
-
             println!("All emails sent (dry run).");
         } else {
-            let client = email_provider::setup_ses_client(matches)?;
-
-            for email in &self.emails {
-                let res = email_provider::send_email(&email, &client);
-                let status = check_send_status(res);
-                println!("{:#?} ... {}", email.receiver, status);
-            }
-
             println!("All emails sent.");
         }
 
@@ -97,7 +86,7 @@ impl BulkEmail {
         let receivers: Vec<String> = self
             .emails
             .iter()
-            .map(|el| el.receiver.to_string())
+            .map(|email| email.receiver.to_string())
             .collect();
         println!(
             "Preparing to send an email to {} recipients: {:#?}",
@@ -127,14 +116,6 @@ impl BulkEmail {
             }
         };
         Ok(confirmation)
-    }
-
-    pub fn archive(&self, matches: &ArgMatches<'_>) -> Result<(), anyhow::Error> {
-        if matches.is_present(arg::MESSAGE_FILE) {
-            MessageTemplate::archive(matches)?;
-        }
-
-        Ok(())
     }
 }
 
@@ -199,18 +180,20 @@ fn create_emails(
     let message = &Message::default(&message_template)?;
 
     let mut emails: Vec<Email> = vec![];
-    let email_series = df_receiver.column(receiver_col)?;
-    let chunked_array = email_series
+    let receiver_series = df_receiver.column(receiver_col)?;
+    let receivers = receiver_series
         .utf8()
         .context("Can't convert series to chunked array")?;
 
-    for email_opt in chunked_array {
-        match email_opt {
-            Some(email) => {
+    for receiver in receivers {
+        match receiver {
+            Some(receiver) => {
+                let mime = Mime::new(matches, sender, receiver, &message)?;
                 emails.push(Email {
                     sender: sender.to_string(),
-                    receiver: email.to_string(),
+                    receiver: receiver.to_string(),
                     message: message.clone(),
+                    mime,
                 });
             }
             None => continue,
@@ -248,11 +231,9 @@ fn create_personalized_emails(
             };
         }
 
-        let receiver_col: &str;
-
         // If argument 'RECEIVER_COLUMN' is not present the default value 'email' will be used
-        match matches.value_of(arg::RECEIVER_COLUMN) {
-            Some(col_name) => receiver_col = col_name,
+        let receiver_col = match matches.value_of(arg::RECEIVER_COLUMN) {
+            Some(receiver_col) => receiver_col,
             None => {
                 return Err(anyhow!(
                     "Missing value for argument '{}'",
@@ -260,7 +241,6 @@ fn create_personalized_emails(
                 ))
             }
         };
-
         let receiver = df_receiver
             .column(receiver_col)
             .context(format!(
@@ -269,13 +249,15 @@ fn create_personalized_emails(
             ))?
             .utf8()?
             .get(i)
-            .unwrap()
+            .context("Can't get value of chunked array")?
             .to_string();
+        let mime = Mime::new(matches, sender, &receiver, &message)?;
 
         emails.push(Email {
             sender: sender.to_string(),
             receiver,
             message,
+            mime,
         });
     }