+++ src/email_builder/email.rs
@@ -1,6 +1,7 @@
+use super::Mime;
 use crate::{
     arg,
-    email_handler::{Confirmed, Message, MessageTemplate},
+    email_builder::{Confirmed, Message},
     email_provider,
     helper::{check_send_status, format_green},
 };
@@ -13,6 +14,7 @@ pub struct Email {
     pub sender: String,
     pub receiver: String,
     pub message: Message,
+    pub mime: Mime,
 }
 
 impl Email {
@@ -23,10 +25,12 @@ impl Email {
         ) {
             (Some(sender), Some(receiver)) => {
                 let message = Message::new(matches)?;
+                let mime = Mime::new(matches, sender, receiver, &message)?;
                 let email = Email {
                     sender: sender.to_string(),
                     receiver: receiver.to_string(),
                     message,
+                    mime,
                 };
                 Ok(email)
             }
@@ -41,23 +45,17 @@ impl Email {
     }
 
     pub fn send(&self, matches: &ArgMatches<'_>) -> Result<(), anyhow::Error> {
-        println!("Sending email to 1 recipient ...");
-        match matches.is_present(arg::DRY_RUN) {
-            true => {
-                // Setup client but do not send email
-                let _client = email_provider::setup_ses_client(matches)?;
-                println!("{:#?} ... {}", self.receiver, format_green("dry run"));
-                println!("All emails sent (dry run).");
-                Ok(())
-            }
-            false => {
-                let client = email_provider::setup_ses_client(matches)?;
-                let res = email_provider::send_email(self, &client);
-                let status = check_send_status(res);
-                println!("{:#?} ... {}", self.receiver, status);
-                println!("All emails sent.");
-                Ok(())
-            }
+        if matches.is_present(arg::DRY_RUN) {
+            // Setup client but do not send email
+            let _client = email_provider::setup_ses_client(matches)?;
+            println!("{:#?} ... {}", self.receiver, format_green("dry run"));
+            Ok(())
+        } else {
+            let client = email_provider::setup_ses_client(matches)?;
+            let res = email_provider::send_raw_email(self, &client);
+            let status = check_send_status(res);
+            println!("{:#?} ... {}", self.receiver, status);
+            Ok(())
         }
     }
 
@@ -92,8 +90,8 @@ impl Email {
     }
 
     pub fn archive(&self, matches: &ArgMatches<'_>) -> Result<(), anyhow::Error> {
-        if matches.is_present(arg::MESSAGE_FILE) {
-            MessageTemplate::archive(matches)?;
+        if matches.is_present(arg::ARCHIVE) {
+            self.mime.archive(matches)?;
         }
 
         Ok(())