+++ CHANGELOG.md
@@ -3,7 +3,12 @@
 ### v0.2.0
 
 - features
+  - Support for arrow2
   - Support `bytea` type in postgres
+  - Implement `--file-type`, `--image-column`, `--image-name`, and `--save-dir` for `query`
+  - Implement `--archive-dir` for `send` and `send-bulk`
+  - Implement MIME format
+  - Implement `--attachment` for `send` and `send-bulk`
 
 ### v0.1.3